Enhance DEBIAN packaging files
Many DEBIAN packaging files have been improved to get rid of warnings 
from lintian.

Tests have been modified to check for the existing python executable 
before attempting to run a test. The check is based on the invocation of 
python itself so that the cache has been warmed up for the actual test 
execution. The attack test is now conditional on whether the user is 
root or not.