Improve bad interpreter state instance detection.
The check for a valid interpreter state has been enhanced with a check 
for a valid thread state instance. Invalid thread state instances could 
be the sign of left over interpreter states that are no longer valid. 
This change was prompted by issue #18 and is a tentative fix for it.

En passant, logging was also improved by flagging error messages that 
are fatal so that they can be better reported to the user.