Improve symbol de-referencing and overhaul wait logic.
The _PyThreadState_Current was incorrectly de-referenced. This has been
fixed by correctly reading a pointer rather than a structure from the
remote address of the symbol.

Added the `interp_head` symbol even though this is usually available
from binaries with debug symbols.

Improved support for _PyRuntimeState de-referencing of interpreter and
current thread. This allows determining which thread holds the GIL or if
the interpreter is idle (NULL current thread).

The wait logic has been simplified and now performs the following checks
in a loop until the interpreter is resolved:

- try to de-reference the interpreter head (either from interp_head or
from _PyRuntime);
- try to get the interpreter state from the current thread, if set;
- scan the BSS directly.

If these steps fail and the heap scan is enabled, then the heap is
scanned as before.