libpython (#20) Add support for embedded Python
* Add support for libpython on Linux.

Tested with Apache2 and mod_wsgi using Python 3.6.7.

* Improve bad interpreter state instance detection.

The check for a valid interpreter state has been enhanced with a check 
for a valid thread state instance. Invalid thread state instances could 
be the sign of left over interpreter states that are no longer valid. 
This change was prompted by issue #18 and is a tentative fix for it.

En passant, logging was also improved by flagging error messages that 
are fatal so that they can be better reported to the user.

* Improve symbol de-referencing and overhaul wait logic.

The _PyThreadState_Current was incorrectly de-referenced. This has been
fixed by correctly reading a pointer rather than a structure from the
remote address of the symbol.

Added the `interp_head` symbol even though this is usually available
from binaries with debug symbols.

Improved support for _PyRuntimeState de-referencing of interpreter and
current thread. This allows determining which thread holds the GIL or if
the interpreter is idle (NULL current thread).

The wait logic has been simplified and now performs the following checks
in a loop until the interpreter is resolved:

- try to de-reference the interpreter head (either from interp_head or
from _PyRuntime);
- try to get the interpreter state from the current thread, if set;
- scan the BSS directly.

If these steps fail and the heap scan is enabled, then the heap is
scanned as before.

* Re-enabled Python 2.3 and 2.4 tests

* Add shared library support for Windows and MacOS

The changes required were small because Python on Windows and MacOS relied on shared objects already.

* Align coding among platforms.

The coding of the VM maximal bounds, version from shared object, and 
symbol loading has been aligned across the three supported platforms.