Add memory profiling capabilities
Using the current thread symbols one can reliably detect which thread 
holds the GIL. With this information at hand one can sample the resident 
memory usage of the process and provide memory usage metrics.

This commit implements the required methods on all the supported 
platforms (only Linux has been tested though, so potentially still a WIP 
for the other two platforms). Memory metrics can be produced with the -m 
switch. The new switch -f allows outputting time and memory metrics for 
each sample.

This commit also introduces support for external output file. When no 
file is specified with the new -o option, the child process STDOUT is 
closed to prevent bad output to interleave with Austin samples. This 
features is on Linux only and needs to be ported to the other two 
platforms.