Add memory profiling capabilities (#21)
Using the current thread symbols one can reliably detect which thread 
holds the GIL. With this information at hand one can sample the resident 
memory usage of the process and provide memory usage metrics.

This commit implements the required methods on all the supported 
platforms (only Linux has been tested though, so potentially still a WIP 
for the other two platforms). Memory metrics can be produced with the -m 
switch. The new switch -f allows outputting time and memory metrics for 
each sample.

This commit also introduces support for external output file. When no 
file is specified with the new -o option, the child process STDOUT is 
closed to prevent bad output to interleave with Austin samples. This 
features is on Linux only and needs to be ported to the other two 
platforms.

* Enhanced command line parsing and parameter handling.

Global variables have been replaced with a single global structure.

Support for output file has been extended to Win and MacOS.
Travis CI configuration also updated to compile with the -lpsai flag
on Windows.

String duplication has been simplified with the use of str(n)dup 
(inspired by #19).

* Enhance tests

The fork test has been enhanced to test memory profiling. The tests
have been sped up and the Python 2.4 fork test has been disabled
because of unreliable output. This will have to be investigated.

* Extend output file option to Windows.

Redirecting to the NULL device instead of closing the STDOUT file 
descriptor for the child process.