Add multi-process support for Linux
Created a new module to handle a list of Python processes that are 
descendant of the given command or the attached process. At every 
iteration, every valid running process in the list is sampled.

The code has been refactored. In particular the sample output generation 
has been moved in py_thread.c and the higher level sampling logic in 
py_proc.c. This allows the new data structure to easily call on these 
methods.

The test suite has been extended to test multi-process support when 
starting a new multi-process Python program.