Test multiprocess support on Windows
Fixed part of the logic to allow the multiprocess support to work on 
Windows.

Added extra process information with the process handle. Previously the 
pid field of py_proc was (ab)used with the process handle. Now it stores 
the actual process ID and the handle has been moved to a field of the 
extra process information structure.