Hotfix/3.8 support (#44)
* hotfix(3.8-support): Fix broken support for Python 3.8

The _PyRuntimeState structure changed between Python 3.7 and 3.8. This hasn't been caught in tests because the fallback on BSS scanning on Linux worked. However, this broke Austin on MacOS and a fix is therefore required.

Enhanced MacOS coding by adding additional delay at startup to ensure that the VM maps have stabilised. When the maps are looked up too soon, the wrong symbol addresses are inferred.

Enhanced multi-processing tests

Fixed support for Python distributions on Mac OS that do not use a shared library

Update Mac OS instructions in the README to mention that sudo is required.

Also, metrics are now logged on STDERR too.