refactor(macos_info): Improve error reporting on MacOS
When trying to launch Austin without sudo on Mac OS, no error
is produced on screen that informs the user that they should
use sudo and ensure they are not invoking a Python binary
installed in a protected location. This change improves this
by showing a more informative error message in these cases.