fix(version): Fix version detection
On Linux, the executable might end up being detected as a Python shared 
object. We try to execute the library binary to retrieve the Python 
version as a first option. If that fails then we fall back on the actual 
binary (if detected) otherwise we parse the library name, as done 
previously.