fix(win): follow single child of non-Python process
On Windows, Python uses a wrapper to launch the actual
Python interpreter process. When in single process mode,
we try to find the single child Python interpreter
process and attach to that instead.

Also added instructions to use an explicit output file since the
wrapper might fail to duplicate the STDOUT handle and prevent the
application from running.