fix _pthread_tid_offset check for older versions of glibc
For recent versions of glic, the pthread struct is like,

{
    ...
    pid_t tid,          // size_of(pid_t) == 4
    pid_t pid_unused,   // value is 0
    ...
}

Before this commit, the _pthread_tid_offset check logic iterates the memory
block in every 8 bytes (uintptr_t), and it works because pid_unused is 0.

However, in older versions of glibc like 2.17 which is used by RHEL/Centos 7,
its pthread struct is like below and comparing 8 bytes it does not match
the tid because the pid field is non-zero.

{
    ...
    pid_t tid,          // sizeof(pid_t) == 4
    pid_t pid,          // value is not 0 but pid
    ...
}