fix(win): support Python launchers
Python launchers duplicate standard streams on child process
creation, which makes it impossible to redirect STDOUT to the NULL
device when starting a command with Austin. To resolve this issue
we now start a reader thread that ensure the STDOUT buffer never
gets full, thus preventing the child from blocking when emitting
output.