refactor(version): make version descriptor local
The use of a global Python version descriptor breaks the possible
support for Python processes that spawn children that use a
different version of the interpreter. By moving all the version
information to the py_proc structure we ensure that each process
can make use of the correct Python version descriptor.