fix: prioritise version inference from file name
This change gives priority to inference of the Python version from
the file name of the Python library binary (if any). This is to
potentially prevent trying to execute the library with the -V
option to get the version. In some cases, this causes the spawned
subprocess to segfault. This can cause misleading terminal output,
but does not affect Austin functionalities.

Resolves #131.