fix: multiprocess fail-fast (#149)
When trying to profile a child process that is not recognised as
Python process, Austin tries to resolve the interpreter state too
many times, causing poor overall performance. This change ensures
that if we fail to attach to a child Process we give up quickly to
allow for samples to be collected efficiently for the processes
that are correctly identified as Python processes.