refactor: drop alternative output format (#155)
The alternative output format was a hack to allow for a quick
visualisation with line breakdown using existing flame graph
visualisation tools. If required, the same effect can be obtained
with ad-hoc tools and therefore it is not necessary to offer this
option from Austin. The benefit is a slight simplification of the
CLI interface.