chore: ignore unused compiler options
When the -Werror option was added, it caused clang to error on the use
of the ignored -s option. This resulted in build failures on macOS for
the conda-forge package (see conda-forge/austin-feedstock#21). This
change adds the -Wno-unused-command-line-argument option to prevent the
compiler from treating unused command line options as errors while still
treating all other warnings as errors.