tests: run uwsgi tests in virtual environment
We run the uwsgi tests in separated virtual environment to ensure that
we install it using the interpreter that Austin is being tested against.
Currently, we install the dependency in the environment used to run
pytest, which means that we test uwsgi with the same Python version in
every job.