fix(linux): improved container support
We make use of the /proc/root of the procfs to improve the chances
that we can infer the required information, such as the location
of the runtime structure and the interpreter version from binaries
pointed at by the actual process root. This is relevant for
processes running inside containers, since their FS root does not
coincide with the host root.