Fix build failure on non-arm 32bit architectures
An incompatible format triggers -Werror=format-security. %zu is widely
available and is the correct type for size_t. The correct format string
to be used for uintptr_t is defined in inttypes.h.

When casting from a pointer to an integer, we trip up
-Werror=pointer-to-int-cast unless the target integer type has exactly
the same size as a pointer.