perf: direct remote access of fields of big structures
Copying over big data structures via kernel space is generally
expensive. We do direct field access instead in these cases.
Currently, the only big structure is the interpreter state, that
needs to be copied over in the sampling loop, for each interpreter.
We change the code to access the fields that we care about directly
instead of copying the whole structure over to user space. Local
tests show a 3x performance improvement when sampling in a tight
loop.