chore: remove latest version fallback
We remove the fallback to the latest version when trying to profile
a version of CPython that is not yet supported. The changes in the
ABI have proven to be very substantial in between the latest minor
releases so the chances of the latest version working with new
CPython releases are very low.