refactor!: replace sleepless option with cpu
We replace the sleepless option with the semantically more correct cpu
option. Originally, the sleepless switch performed heuristics to
filter out frames that were likely idle. Over time Austin has developed
a better way of determining if a stack is on-CPU and has a full-fledged
CPU mode. We make this change to reflect the current state.