perf: page size cap
We introduce a page size cap for remote VM reads that can be
configured via environment variables. A smaller page size on
systems that have large defaults (e.g. 16K on M1) appears to be
beneficial in terms of performance. We expose this via env var to
allow tuning it if required.