chore: event buffering in pipe mode
When in pipe mode we do event buffering. This is the equivalent of
line buffering but for the two main Austin events, that is samples
and metadata. With this change, any tool communicating with Austin
via a pipe will get the bytes for each event immediately instead
of at the default stream buffer size. This allows tools to get
events from Austin in real-time as they are collected instead of
periodic bursts.