refactor: terminal output
We refactor the code around terminal output. We make the code more
readable by using macros for the ANSI escape sequences. We also
re-arrange the layout of some of the terminal output, such as the
sampling statistics to convey more user-relevant information. We
also include some extra metadata that can be useful for tools that
will process the output after collection.